refactor: 기존 BlackList 기능을 수행하는 CustomerRepository 를 BlackListRepository 로 변경
