+++ src/main/java/study/dev/spring/customer/application/BlackListRepository.java
@@ -1,0 +1,10 @@
+package study.dev.spring.customer.application;
+
+import java.util.List;
+
+import study.dev.spring.customer.application.dto.CustomerData;
+
+public interface BlackListRepository {
+
+	List<CustomerData> findAll();
+}
