+++ src/test/java/study/dev/spring/customer/application/BlackListRepositoryTest.java
@@ -1,4 +1,4 @@
-package study.dev.spring.customer.infrastructure;
+package study.dev.spring.customer.application;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -8,12 +8,13 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import study.dev.spring.common.stub.FileUtilsStub;
-import study.dev.spring.customer.domain.Customer;
+import study.dev.spring.customer.application.dto.CustomerData;
+import study.dev.spring.customer.infrastructure.FileBlackListRepository;
 
-@DisplayName("[FileCustomerRepository Test] - Infrastructure Layer")
-class FileCustomerRepositoryTest {
+@DisplayName("[BlackListRepository 테스트] - Application Layer")
+class BlackListRepositoryTest {
 
-	private final FileCustomerRepository fileCustomerRepository = new FileCustomerRepository(
+	private final BlackListRepository fileCustomerRepository = new FileBlackListRepository(
 		List.of(new FileUtilsStub()), "path"
 	);
 
@@ -21,7 +22,7 @@ class FileCustomerRepositoryTest {
 	@DisplayName("[모든 유저를 조회한다]")
 	void findAllTest() {
 		//when
-		List<Customer> actual = fileCustomerRepository.findAll();
+		List<CustomerData> actual = fileCustomerRepository.findAll();
 
 		//then -> save 로직이 없기 때문에 비어있는지만 검사
 		assertThat(actual).isEmpty();
